+++ buckaroo/analysis_management.py
@@ -1,7 +1,49 @@
+import numpy as np
+import pandas as pd
+
 from buckaroo.pluggable_analysis_framework import (
-    ColAnalysis, order_analysis, check_solvable, NotProvidedException,
-    produce_summary_df)
+    ColAnalysis, order_analysis, check_solvable, NotProvidedException)
+
+
 FAST_SUMMARY_WHEN_GREATER = 1_000_000
+
+
+def produce_summary_df(df, ordered_objs, df_name='test_df'):
+    """
+    takes a dataframe and a list of analyses that have been ordered by a graph sort,
+    then it produces a summary dataframe
+    """
+    errs = {}
+    summary_col_dict = {}
+    table_hint_col_dict = {}
+    #figure out how to add in "index"... but just for table_hints
+    for ser_name in df.columns:
+        ser = df[ser_name]
+        summary_ser = pd.Series({}, dtype='object')
+        table_hint_dict = {}
+        for a_kls in ordered_objs:
+            try:
+                summary_res = a_kls.summary(ser, summary_ser, ser)
+                for k,v in summary_res.items():
+                    summary_ser.loc[k] = v
+                for k,v in a_kls.table_hints(sampled_ser, summary_ser, table_hint_dict):
+                    table_hint_dict[k] = v
+            except Exception as e:
+                print("summary_ser", summary_ser)
+                errs[ser_name] = e, a_kls
+                continue
+        summary_col_dict[ser_name] = summary_ser
+        table_hint_col_dict[ser_name] = table_hint_dict
+    if errs:
+        for ser_name, err_kls in errs.items():
+            err, kls = err_kls
+            print("%r failed on %s with %r" % (kls, ser_name, err))
+        print("Reproduce")
+        print("from pluggable_analysis import test_ser")
+        for ser_name, err_kls in errs.items():
+            err, kls = err_kls
+            print("%s.summary(test_ser.%s)" % (kls.__name__, ser_name))
+    return pd.DataFrame(summary_col_dict), table_hint_col_dict
 
 
 class AnalsysisPipeline(object):
@@ -31,9 +73,9 @@ class AnalsysisPipeline(object):
         """
         pass
 
-    def produce_summary_df(self, input_df):
-        output_df = produce_summary_df(input_df, self.ordered_a_objs)
-        return output_df
+    def process_df(self, input_df):
+        output_df, table_hint_dict = produce_summary_df(input_df, self.ordered_a_objs)
+        return output_df, table_hint_dict
 
     def add_analysis(self, new_aobj):
         new_cname = new_aobj.cname()
@@ -54,7 +96,9 @@ class DfStats(object):
         self.df = self.get_operating_df(df, force_full_eval=False)
         self.col_order = self.df.columns
         self.ap = AnalsysisPipeline(col_analysis_objs)
-        self.sdf = self.ap.produce_summary_df(self.df)
+        self.sdf, self.table_hints = self.ap.process_df(self.df)
+
+
 
     def get_operating_df(self, df, force_full_eval):
         rows = len(df)
@@ -65,3 +109,5 @@ class DfStats(object):
             return df.sample(np.min([50_000, len(df)]))
         else:
             return df
+
+
