+++ tests/pluggable_analysis_framework_test.py
@@ -4,10 +4,10 @@ import pandas as pd
 import numpy as np
 import graphlib
 from buckaroo.pluggable_analysis_framework import (
-    ColAnalysis, order_analysis, check_solvable, NotProvidedException,
-    produce_summary_df)
+    ColAnalysis, order_analysis, check_solvable, NotProvidedException)
+
 from buckaroo.analysis_management import (
-    AnalsysisPipeline)
+    AnalsysisPipeline, produce_summary_df)
 
 from buckaroo.analysis import (TypingStats, DefaultSummaryStats)
 
@@ -120,7 +120,7 @@ class TestAnalysisPipeline(unittest.TestCase):
     def test_pipeline_base(self):
         ap = AnalsysisPipeline([TypingStats, DefaultSummaryStats])
         #just verify that there are no errors
-        ap.produce_summary_df(df)
+        ap.process_df(df)
 
     def test_add_aobj(self):
         ap = AnalsysisPipeline([TypingStats, DefaultSummaryStats])
@@ -133,7 +133,7 @@ class TestAnalysisPipeline(unittest.TestCase):
             def summary(sampled_ser, summary_ser, ser):
                 return dict(foo=8)
         ap.add_analysis(Foo)
-        sdf = ap.produce_summary_df(df)
+        sdf, _unused = ap.process_df(df)
         self.assertEqual(sdf.loc['foo']['tripduration'], 8)
 
     def test_replace_aobj(self):
@@ -147,7 +147,7 @@ class TestAnalysisPipeline(unittest.TestCase):
             def summary(sampled_ser, summary_ser, ser):
                 return dict(foo=8)
         ap.add_analysis(Foo)
-        sdf = ap.produce_summary_df(df)
+        sdf, _unused = ap.process_df(df)
         self.assertEqual(sdf.loc['foo']['tripduration'], 8)
         #18 facts returned about tripduration
         self.assertEqual(len(sdf['tripduration']), 18)
@@ -161,7 +161,7 @@ class TestAnalysisPipeline(unittest.TestCase):
             def summary(sampled_ser, summary_ser, ser):
                 return dict(foo=9)
         ap.add_analysis(Foo)
-        sdf2 = ap.produce_summary_df(df)
+        sdf2, _unused = ap.process_df(df)
         self.assertEqual(sdf2.loc['foo']['tripduration'], 9)
         #still 18 facts returned about tripduration
         self.assertEqual(len(sdf2['tripduration']), 18)
